[vm] Fix to ensure positions from CSMs in precompiled are as expected.
Token positions in the CodeSourceMap are converted to line
numbers (or kNoSourcePos if no line number) in precompiled mode.
Fix the reader so that it uses kNoSource as the initial position
in precompiled mode instead of kDartPrologue.

This conversion never happens to token positions in PcDescriptors
objects, so we should never treat a position coming from them (i.e.,
via GetTokenIndexForPC) as a line.

Bug: https://github.com/dart-lang/sdk/issues/44492

TEST=Run on debug precomp trybot to verify fixed.

Cq-Include-Trybots: luci.dart.try:vm-kernel-precomp-linux-debug-x64-try
Change-Id: Ib4d4be94240997b8bcd6b134b786e82d77875a21
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/176246
Commit-Queue: Tess Strickland <sstrickl@google.com>
Reviewed-by: Martin Kustermann <kustermann@google.com>