[ddc] Fix equals operator on native types
Emit `===` instead of `[$equals]` when the LHS is represented as a
primitive type.

The calculation of the left type was resulting in Object instead of a
more precise type. This caused the check for primitives types to always
fail.

Provides 20%+ performance improvements on the Wordsolver and
CollectionSieves-removeWhere benchmarks.
https://golem.corp.goog/Comparison?repository=dart&team=dartdevc#targetA%3Ddartdevc%3BmachineTypeA%3Dlinux-x64%3BrevisionA%3D89773%3BpatchA%3Dnshahan-fix-primative-equals%3BtargetB%3Ddartdevc%3BmachineTypeB%3Dlinux-x64%3BrevisionB%3D89771%3BpatchB%3DNone

Change-Id: Ica980f28bf8df627571b9aa53e77478e2b1de17f
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/176069
Reviewed-by: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Mark Zhou <markzipan@google.com>
Commit-Queue: Nicholas Shahan <nshahan@google.com>