[dart2js] delete bound_closure_super_test
This test has been failing for a long time because it had invalid expectations.

The test was written to test that, when overriding a member and making a
super call on a native class, we properly call the super Dart stub and
the super JavaScript method.

The former was handled properly (in the test the default parameter value
was injected according to the super type), but the latter was never
implemented (the generated code invoked `receiver.foo` instead of
`BB.prototype.foo.call(receiver, ...)`

We don't have plans to add support for the latter at this time. In the
future we do want to consider adding a static error checker to ensure
that we don't attempt to use this missing feautre.

Change-Id: I5f5179ebdf31d24d4a0b1edb8f947b94f85934ba
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/178060
Commit-Queue: Sigmund Cherem <sigmund@google.com>
Reviewed-by: Stephen Adams <sra@google.com>