Change FlowAnalysisDebug to wrap FlowAnalysis.
Previously, it wrapped _FlowAnalysisImpl.  I'm planning a series of
changes to allow legacy type promotion to be expressed using the same
API as flow analysis; in order to make this possible to debug using
FlowAnalysisDebug, it needs to be able to wrap any implementation of
FlowAnalysis, not just _FlowAnalysisImpl.

Change-Id: Ic1ae87c22de6d4ff1e0d5d86090726272f741c34
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/179380
Reviewed-by: Konstantin Shcheglov <scheglov@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>