[web] partial migration of native tests to support dartdevc
Most files are changed in two ways:

* add `self.` so that the code can run with "use strict" (which ddc
does by default when generating code)

* call the js-helper that installs the interceptor/type-extensions in
DDC on each test.

This was mostly done mechanically and gets about 50% of the tests under
dart2js/native running and passing in dartdevc.  There are, however,
many tests here that are dart2js-specific or for features we don't want
to prioritize at this time (like custom elements), so after this the
next step should probably be to reoarganize the tests and split them in
general native support vs dart2js-specific.

Change-Id: I37759d854f9f51c185789471454d37b6bc78af90
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/177821
Reviewed-by: Stephen Adams <sra@google.com>
Reviewed-by: Nicholas Shahan <nshahan@google.com>