[web] fix regression on two native tests.
Recent changes to support native tests in DDC broke 2 tests only in chrome.
These broke because exporting the constructor changed the intent of these tests
in the first place.

This wasn't detected in d8 because we are using a mix of `self` and `window`,
and only in the browser they mapped to the same object.

Change-Id: I795b59ae2c0f34fdd7fc1730fbc3097dae2fee75
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/179660
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>